Integrate Marketplace With Backend #61 reintroduce the usage of the getProdList() method and callback to fetch products data from the database and create a new version of ProductService.js to connect with the backend. This will allow the product list to be fetched from the database on the backend.
